pyproject.toml
[tool.poetry]
name = "movies-notifications-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
fastapi = "0.95.1"
uvicorn = "0.22.0"
gunicorn = "20.1.0"
fastapi-admin = "1.0.4"
redis = "4.5.5"
asyncpg = "0.27.0"
types-redis = "4.5.5.2"
yoyo-migrations = "^8.2.0"
dependency-injector = "^4.41.0"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
# tests
pytest = "7.3.1"
pytest-asyncio = "0.21.0"
aiohttp = "3.8.4"
httpx = "0.24.0"
faker = "18.9.0"
# linters
black = "23.3.0"
isort = "5.12.0"
flake8 = "6.0.0"
mypy = "1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    .venv
    | venv
    | migrations
    | rabbitmq
)/
'''

[tool.isort]
line_length = 79
combine_as_imports = true
include_trailing_comma = true
lines_after_imports = 2
skip = [".venv", "venv", "migrations", "rabbitmq"]

[tool.mypy]
exclude = ['tests','docs','nginx','notifications_db']
ignore_missing_imports = true

